Fix `FOREACH x IN NULL` bug
Negative testing of the new procedure which uses JSON-based mapping
details revealed that when a nonexistent json_type was passed in (e.g.
"ample_json") the SELECT statement used to create an array of mapped
schema names returned NULL rather than an empty array. This in turn
caused the FOREACH statement to fail.

This bug does not affect the original, "table-based mapping details"
version of the procedure which iterates over a recordset, not an array.
The new procedure has been amended to follow that pattern.