BREAKING: Add 'OR' conditions to query statements
For the deepest nested groups, combine the query statements obtained in
the last commit to produce alternative conditions for each element in an
array.

The existing test data does not have multiple arrays in a single object
so does not fully test the new code and should be updated.

A defect was noticed where 'null' values are not being handled correctly
- this must be resolved before release.