Fix null handling
If a JSON field has null value, it will always be included in the query.

If a JSON field is missing, it will be checked for a null value ONLY IF
it is linked to another field which is present in the JSON, resulting in
a condition like "foo = 6 AND bar IS NULL".

If all mapped fields are missing for a JSON object, it will not produce
a query, avoiding the incorrect "foo IS NULL and bar IS NULL" conditions
which were previously generated.

There may be edge cases where deeply nested fields with null values, not
associated with any non-null fields, are lost from the final query due
to how the loop is set up - this is a very unlikely scenario and the fix
is non-trivial so will not be explored further.